chore: move `TransformResult` to types package
