+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Chore & Maintenance
 
+- `[@jest/test-result]` Remove dependency on `@jest/transform`, which lead to a sprawling dependency tree ([#9747](https://github.com/facebook/jest/pull/9747))
 - `[@jest/transform]` Expose type `TransformedSource` ([#9736](https://github.com/facebook/jest/pull/9736))
 
 ### Performance
