+++ packages/jest-test-result/package.json
@@ -18,7 +18,6 @@
   },
   "dependencies": {
     "@jest/console": "^25.2.3",
-    "@jest/transform": "^25.2.4",
     "@jest/types": "^25.2.3",
     "@types/istanbul-lib-coverage": "^2.0.0",
     "collect-v8-coverage": "^1.0.0"
