+++ packages/jest-transform/src/types.ts
@@ -6,7 +6,7 @@
  */
 
 import type {RawSourceMap} from 'source-map';
-import type {Config} from '@jest/types';
+import type {Config, TransformTypes} from '@jest/types';
 
 export type ShouldInstrumentOptions = Pick<
   Config.GlobalConfig,
@@ -36,12 +36,7 @@ export type TransformedSource =
   | {code: string; map?: FixedRawSourceMap | string | null}
   | string;
 
-export type TransformResult = {
-  code: string;
-  originalCode: string;
-  mapCoverage: boolean;
-  sourceMapPath: string | null;
-};
+export type TransformResult = TransformTypes.TransformResult;
 
 export type TransformOptions = {
   instrument: boolean;
